@@ -1,13 +1,14 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 28
+    compileSdk 30
     defaultConfig {
         applicationId "org.ajcm.hiad"
-        minSdkVersion 14
-        targetSdkVersion 28
-        versionCode 31
-        versionName "1.9.35"
+        minSdkVersion 19
+        targetSdk 30
+        versionName "1.9.4"
+        versionCode 34
+        multiDexEnabled true
     }
     buildTypes {
         release {
@@ -15,14 +16,16 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+    }
 }
 
 dependencies {
-    def VERSION_ANDROID_SUPPORT = '28.0.0'
-
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation 'androidx.legacy:legacy-support-v4:1.0.0'
-    implementation 'androidx.appcompat:appcompat:1.1.0'
+    implementation 'androidx.appcompat:appcompat:1.3.1'
     implementation 'com.google.android.material:material:1.0.0'
 
     implementation "com.google.firebase:firebase-ads:17.1.3"
@@ -33,15 +36,17 @@ dependencies {
 
     implementation 'com.sothree.slidinguppanel:library:3.2.0'
     implementation 'fr.avianey.com.viewpagerindicator:library:2.4.1.1@aar'
-    implementation 'org.zakariya.stickyheaders:stickyheaders:0.7.6@aar'
+    implementation 'org.zakariya.stickyheaders:stickyheaders:0.7.10@aar'
     implementation 'com.github.ivbaranov:materialfavoritebutton:0.1.4'
     implementation 'com.github.chrisbanes:PhotoView:2.0.0'
 
-    implementation 'com.github.bumptech.glide:glide:4.10.0'
+    implementation 'com.github.bumptech.glide:glide:4.11.0'
     annotationProcessor 'com.github.bumptech.glide:compiler:4.10.0'
 
-    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
-    testImplementation 'junit:junit:4.12'
+    implementation 'androidx.constraintlayout:constraintlayout:2.1.2'
+    testImplementation 'junit:junit:4.13.2'
+
+    implementation "androidx.multidex:multidex:2.0.1"
 }
 
 apply plugin: 'com.google.gms.google-services'
