@@ -6,11 +6,12 @@ import android.database.Cursor;
 import android.database.SQLException;
 import android.database.sqlite.SQLiteDatabase;
 import android.util.Log;
-import android.util.Pair;
 
 import org.ajcm.hiad.models.Himno;
 import org.ajcm.hiad.models.Himno1962;
 import org.ajcm.hiad.models.Himno2008;
+import org.ajcm.hiad.models.Historial;
+import org.ajcm.hiad.utils.StringUtils;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -24,6 +25,7 @@ public class DBAdapter {
     public static final String DATABASE_TABLE_2008 = "himnario2008";
     public static final String DATABASE_TABLE_1962 = "himnario1962";
     public static final String DATABASE_TABLE_CATEGORIA = "categorias";
+    public static final String DATABASE_TABLE_HISTORIAL = "historial";
     private final Context context;
     private DatabaseHelper DBHelper;
     private SQLiteDatabase db;
@@ -87,7 +89,6 @@ public class DBAdapter {
         while (query.moveToNext()) {
             himnos.add(query.getInt(Himno2008.Columns.numero.ordinal()));
         }
-
         query.close();
         return himnos;
     }
@@ -97,6 +98,34 @@ public class DBAdapter {
         ContentValues values = new ContentValues();
         values.put(Himno2008.Columns.favorito.name(), fav);
         db.update(tableVersion(version2008), values, Himno2008.Columns.numero.name() + " = " + numero, null);
+    }
+
+    public void setHistorial(int numero) {
+        open();
+        Log.e(TAG, "setHistorial: open");
+        Cursor query = db.query(DATABASE_TABLE_HISTORIAL,
+                null,
+                Historial.Columns.numero.name() + " = " + numero,
+                null, null, null, null, null);
+        int frecuency = 1;
+        boolean newHistory = true;
+        if (query.moveToFirst()) {
+            newHistory = false;
+            frecuency = query.getInt(query.getColumnIndex(Historial.Columns.frecuencia.name())) + 1;
+        }
+        Log.e(TAG, "setHistorial: frecuency " + frecuency);
+        ContentValues values = new ContentValues();
+        values.put(Historial.Columns.numero.name(), numero);
+        values.put(Historial.Columns.frecuencia.name(), frecuency);
+        values.put(Historial.Columns.ultima_fecha.name(), StringUtils.getCurrentDateString());
+        Log.e(TAG, "setHistorial: new " + newHistory);
+        if (newHistory) {
+            Log.e(TAG, "setHistorial: save");
+            db.insert(DATABASE_TABLE_HISTORIAL, null, values);
+        } else {
+            Log.e(TAG, "setHistorial: update");
+            db.update(DATABASE_TABLE_HISTORIAL, values, Historial.Columns.numero.name() + " = " + numero, null);
+        }
     }
 
     public void setDuration(int numero, String duration, boolean version2008) {
@@ -136,6 +165,20 @@ public class DBAdapter {
         String titulo = query.getString(1);
         query.close();
         return titulo;
+    }
+
+    public ArrayList<Historial> getHistorial() {
+        ArrayList<Historial> historials = new ArrayList<>();
+        open();
+        Cursor query = db.query(DATABASE_TABLE_HISTORIAL,
+                null,
+                null,
+                null, null, null, Historial.Columns.frecuencia + " DESC", null);
+        while (query.moveToNext()) {
+            historials.add(Historial.fromCursor(query));
+        }
+        query.close();
+        return historials;
     }
 
     private void open() throws SQLException {
